crawlproject/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0

# Kafka dependencies
kafka-python>=2.0.2

# Spark dependencies
pyspark>=3.5.0

# Data processing
scikit-learn>=1.1.0
scipy>=1.9.0

# Utilities
python-dateutil>=2.8.0
pytz>=2022.1

# Logging and monitoring
structlog>=22.1.0

# HTTP retry and session management
urllib3>=1.26.0

# Development dependencies (optional)
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Testing dependencies
pytest>=7.0.0
pytest-timeout>=2.1.0

# Code quality
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
